feat: introduce conversation management in chat clients
- Added `Conversation` and `AsyncConversation` classes to manage stateful chat interactions.
- Updated `RagbitsChatClient` and `AsyncRagbitsChatClient` to utilize the new conversation classes for handling message history and server state.
- Refactored client methods to improve clarity and maintainability, ensuring a stateless interface for message sending.
- Enhanced the base classes to support conversation management, promoting better organization and usability in chat interactions.
