+++ packages/ragbits-chat/src/ragbits/chat/clients/__init__.py
@@ -1,9 +1,12 @@
 from __future__ import annotations
 
 from .async_client import AsyncRagbitsChatClient
+from .conversation import AsyncConversation, Conversation
 from .sync_client import RagbitsChatClient
 
 __all__ = [
+    "AsyncConversation",
     "AsyncRagbitsChatClient",
+    "Conversation",
     "RagbitsChatClient",
 ]
