+++ packages/ragbits-chat/src/ragbits/chat/clients/async_client.py
@@ -1,20 +1,11 @@
 from __future__ import annotations
 
-from collections.abc import AsyncGenerator
 from typing import Any
 
 import httpx
 
-from .._utils import build_api_url, parse_sse_line
-from ..interface.types import (
-    ChatResponse,
-    ChatResponseType,
-    Message,
-    MessageRole,
-    StateUpdate,
-)
-from .base import AsyncChatClientBase
-from .exceptions import ChatClientRequestError, ChatClientResponseError
+from .base import AsyncChatClientBase, AsyncConversationBase
+from .conversation import AsyncConversation
 
 __all__ = ["AsyncRagbitsChatClient"]
 
@@ -25,7 +16,7 @@ _DEFAULT_HEADERS = {
 
 
 class AsyncRagbitsChatClient(AsyncChatClientBase):
-    """Asynchronous Ragbits chat client."""
+    """Stateless **asynchronous** Ragbits chat client."""
 
     def __init__(
         self,
@@ -36,17 +27,13 @@ class AsyncRagbitsChatClient(AsyncChatClientBase):
         self._base_url = base_url.rstrip("/")
         self._client = httpx.AsyncClient(timeout=timeout, headers=_DEFAULT_HEADERS)
 
-        self.history: list[Message] = []
-        self.conversation_id: str | None = None
-        self.server_state: StateUpdate | None = None
+    def new_conversation(self) -> AsyncConversationBase:
+        """Return a brand-new :class:`AsyncConversation`."""
+        return AsyncConversation(base_url=self._base_url, http_client=self._client)
 
-        self._streaming_response: httpx.Response | None = None
-
-    def new_conversation(self) -> None:
-        """Start a fresh conversation, resetting local state."""
-        self.history.clear()
-        self.conversation_id = None
-        self.server_state = None
+    async def aclose(self) -> None:
+        """Close the underlying *httpx.AsyncClient* session."""
+        await self._client.aclose()
 
     async def ask(
         self,
@@ -54,87 +41,17 @@ class AsyncRagbitsChatClient(AsyncChatClientBase):
         *,
         context: dict[str, Any] | None = None,
     ) -> str:
-        """Send *message* and return the final assistant reply."""
-        parts: list[str] = []
-        async for chunk in self.send_message(message, context=context):
-            if chunk.type is ChatResponseType.TEXT:
-                parts.append(str(chunk.content))
-        return "".join(parts)
+        """Convenience wrapper that proxies through a fresh AsyncConversation."""
+        conv = self.new_conversation()
+        return await conv.ask(message, context=context)
 
     async def send_message(
         self,
         message: str,
         *,
         context: dict[str, Any] | None = None,
-    ) -> AsyncGenerator[ChatResponse, None]:
-        """Asynchronous version of :py:meth:`SyncChatClientBase.send_message`."""
-        user_msg = Message(role=MessageRole.USER, content=message)
-        self.history.append(user_msg)
-        assistant_reply = Message(role=MessageRole.ASSISTANT, content="")
-        self.history.append(assistant_reply)
-        assistant_index = len(self.history) - 1
-
-        merged_context: dict[str, Any] = {}
-        if self.server_state is not None:
-            merged_context.update(self.server_state.model_dump())
-        if self.conversation_id is not None:
-            merged_context["conversation_id"] = self.conversation_id
-        if context:
-            merged_context.update(context)
-
-        payload: dict[str, Any] = {
-            "message": message,
-            "history": [m.model_dump() for m in self.history if m.role is not MessageRole.SYSTEM],
-            "context": merged_context,
-        }
-
-        url = build_api_url(self._base_url, "/api/chat")
-
-        try:
-            async with self._client.stream("POST", url, json=payload) as resp:
-                try:
-                    resp.raise_for_status()
-                except httpx.HTTPStatusError as exc:
-                    raise ChatClientResponseError(
-                        f"Unexpected response from {url}: {exc.response.status_code}"
-                    ) from exc
-
-                self._streaming_response = resp
-
-                async for raw_line in resp.aiter_lines():
-                    if not raw_line:
-                        continue
-                    parsed = parse_sse_line(raw_line)
-                    if parsed is None:
-                        continue
-                    await self._process_incoming(parsed, assistant_index)
-                    yield parsed
-        except httpx.RequestError as exc:
-            raise ChatClientRequestError(f"Error communicating with {url}: {exc}") from exc
-        finally:
-            self._streaming_response = None
-
-    async def stop(self) -> None:
-        """Abort currently running stream (if any)."""
-        if self._streaming_response is not None and not self._streaming_response.is_closed:
-            await self._streaming_response.aclose()
-            self._streaming_response = None
-
-    async def aclose(self) -> None:
-        """Close the underlying *httpx.AsyncClient* session."""
-        await self._client.aclose()
-
-    async def _process_incoming(self, resp: ChatResponse, assistant_index: int) -> None:
-        """Update local client-side state based on *resp*."""
-        if resp.type is ChatResponseType.STATE_UPDATE:
-            self.server_state = resp.content  # type: ignore[assignment]
-        elif resp.type is ChatResponseType.CONVERSATION_ID:
-            self.conversation_id = resp.content  # type: ignore[assignment]
-        elif resp.type is ChatResponseType.MESSAGE_ID:
-            pass
-        elif resp.type is ChatResponseType.TEXT:
-            assistant_msg = self.history[assistant_index]
-            if isinstance(resp.content, str):
-                assistant_msg.content += resp.content
-        elif resp.type is ChatResponseType.REFERENCE:
-            pass
+    ) -> Any:
+        """Stateless proxy to :class:`AsyncConversation.send_message`."""
+        conv = self.new_conversation()
+        async for chunk in conv.send_message(message, context=context):
+            yield chunk
