+++ packages/ragbits-chat/src/ragbits/chat/clients/base.py
@@ -8,7 +8,9 @@ from ..interface.types import ChatResponse
 
 __all__ = [
     "AsyncChatClientBase",
+    "AsyncConversationBase",
     "SyncChatClientBase",
+    "SyncConversationBase",
 ]
 
 
@@ -16,8 +18,8 @@ class SyncChatClientBase(ABC):
     """Abstract base class for synchronous chat clients."""
 
     @abstractmethod
-    def new_conversation(self) -> None:
-        """Start a fresh conversation, resetting local state."""
+    def new_conversation(self) -> SyncConversationBase:
+        """Create and return a new conversation instance."""
 
     @abstractmethod
     def ask(
@@ -46,8 +48,8 @@ class AsyncChatClientBase(ABC):
     """Abstract base class for asynchronous chat clients."""
 
     @abstractmethod
-    def new_conversation(self) -> None:
-        """Start a fresh conversation, resetting local state."""
+    def new_conversation(self) -> AsyncConversationBase:
+        """Create and return a new conversation instance."""
 
     @abstractmethod
     async def ask(
@@ -67,6 +69,60 @@ class AsyncChatClientBase(ABC):
     ) -> AsyncGenerator[ChatResponse, None]:
         """Send *message* and yield streaming :class:`ChatResponse` chunks."""
 
+
+class SyncConversationBase(ABC):
+    """Abstract base class for **synchronous** chat conversations.
+
+    Concrete conversation implementations (e.g. :class:`ragbits.chat.clients.conversation.Conversation`)
+    should inherit from this class to ensure a consistent public interface.  The
+    API mirrors :class:`SyncChatClientBase` but represents a single, *stateful*
+    conversation rather than a stateless client.
+    """
+
+    @abstractmethod
+    def ask(
+        self,
+        message: str,
+        *,
+        context: dict[str, Any] | None = None,
+    ) -> str:
+        """Send *message* and return the assistant reply as a fully-formed string."""
+
+    @abstractmethod
+    def send_message(
+        self,
+        message: str,
+        *,
+        context: dict[str, Any] | None = None,
+    ) -> Generator[ChatResponse, None, None]:
+        """Send *message* and yield streaming :class:`ChatResponse` chunks."""
+
+    @abstractmethod
+    def stop(self) -> None:
+        """Abort a currently running stream (if any)."""
+
+
+class AsyncConversationBase(ABC):
+    """Abstract base class for **asynchronous** chat conversations."""
+
+    @abstractmethod
+    async def ask(
+        self,
+        message: str,
+        *,
+        context: dict[str, Any] | None = None,
+    ) -> str:
+        """Send *message* and return the assistant reply as a fully-formed string."""
+
+    @abstractmethod
+    async def send_message(
+        self,
+        message: str,
+        *,
+        context: dict[str, Any] | None = None,
+    ) -> AsyncGenerator[ChatResponse, None]:
+        """Send *message* and yield streaming :class:`ChatResponse` chunks."""
+
     @abstractmethod
     async def stop(self) -> None:
-        """Abort a currently running request (if any)."""
+        """Abort a currently running stream (if any)."""
