+++ packages/ragbits-chat/src/ragbits/chat/clients/sync_client.py
@@ -5,10 +5,9 @@ from typing import Any
 
 import httpx
 
-from .._utils import build_api_url, parse_sse_line
-from ..interface.types import ChatResponse, ChatResponseType, Message, MessageRole, StateUpdate
-from .base import SyncChatClientBase
-from .exceptions import ChatClientRequestError, ChatClientResponseError
+from ..interface.types import ChatResponse
+from .base import SyncChatClientBase, SyncConversationBase
+from .conversation import Conversation
 
 __all__ = ["RagbitsChatClient"]
 
@@ -19,7 +18,13 @@ _DEFAULT_HEADERS = {
 
 
 class RagbitsChatClient(SyncChatClientBase):
-    """Synchronous Ragbits chat client."""
+    """Stateless *synchronous* Ragbits chat client.
+
+    The sole responsibility of this class is to spawn new
+    :class:`ragbits.chat.clients.conversation.Conversation` objects. All
+    conversation-specific state (history, server-state, etc.) lives in the
+    returned *Conversation* instance.
+    """
 
     def __init__(
         self,
@@ -30,17 +35,13 @@ class RagbitsChatClient(SyncChatClientBase):
         self._base_url = base_url.rstrip("/")
         self._client = httpx.Client(timeout=timeout, headers=_DEFAULT_HEADERS)
 
-        self.history: list[Message] = []
-        self.conversation_id: str | None = None
-        self.server_state: StateUpdate | None = None
+    def new_conversation(self) -> SyncConversationBase:
+        """Return a brand-new :class:`Conversation`."""
+        return Conversation(base_url=self._base_url, http_client=self._client)
 
-        self._streaming_response: httpx.Response | None = None
-
-    def new_conversation(self) -> None:
-        """Start a fresh conversation, resetting local state."""
-        self.history.clear()
-        self.conversation_id = None
-        self.server_state = None
+    def close(self) -> None:
+        """Close the underlying *httpx.Client* session."""
+        self._client.close()
 
     def ask(
         self,
@@ -48,12 +49,15 @@ class RagbitsChatClient(SyncChatClientBase):
         *,
         context: dict[str, Any] | None = None,
     ) -> str:
-        """Send *message* and return the final assistant reply."""
-        assistant_reply_parts: list[str] = []
-        for chunk in self.send_message(message, context=context):
-            if chunk.type is ChatResponseType.TEXT:
-                assistant_reply_parts.append(str(chunk.content))
-        return "".join(assistant_reply_parts)
+        """Convenience wrapper that proxies through a fresh Conversation.
+
+        This preserves the original *ask* method for backwards–compatibility
+        while encouraging users to work with :pyclass:`Conversation` objects
+        directly. Each invocation operates on a **new** conversation so no
+        state is preserved across calls.
+        """
+        conv = self.new_conversation()
+        return conv.ask(message, context=context)
 
     def send_message(
         self,
@@ -61,72 +65,6 @@ class RagbitsChatClient(SyncChatClientBase):
         *,
         context: dict[str, Any] | None = None,
     ) -> Generator[ChatResponse, None, None]:
-        """Send *message* and yield streaming :class:`ChatResponse` chunks."""
-        user_msg = Message(role=MessageRole.USER, content=message)
-        self.history.append(user_msg)
-
-        assistant_reply = Message(role=MessageRole.ASSISTANT, content="")
-        self.history.append(assistant_reply)
-        assistant_index = len(self.history) - 1
-
-        merged_context: dict[str, Any] = {}
-        if self.server_state is not None:
-            merged_context.update(self.server_state.model_dump())
-        if self.conversation_id is not None:
-            merged_context["conversation_id"] = self.conversation_id
-        if context:
-            merged_context.update(context)
-
-        payload: dict[str, Any] = {
-            "message": message,
-            "history": [m.model_dump() for m in self.history if m.role is not MessageRole.SYSTEM],
-            "context": merged_context,
-        }
-
-        url = build_api_url(self._base_url, "/api/chat")
-
-        try:
-            with self._client.stream("POST", url, json=payload) as resp:
-                try:
-                    resp.raise_for_status()
-                except httpx.HTTPStatusError as exc:
-                    raise ChatClientResponseError(
-                        f"Unexpected response from {url}: {exc.response.status_code}"
-                    ) from exc
-
-                self._streaming_response = resp
-
-                for raw_line in resp.iter_lines():
-                    if not raw_line:
-                        continue
-                    parsed = parse_sse_line(raw_line)
-                    if parsed is None:
-                        continue
-
-                    self._process_incoming(parsed, assistant_index)
-                    yield parsed
-        except httpx.RequestError as exc:
-            raise ChatClientRequestError(f"Error communicating with {url}: {exc}") from exc
-        finally:
-            self._streaming_response = None
-
-    def stop(self) -> None:
-        """Abort currently running stream (if any)."""
-        if self._streaming_response is not None and not self._streaming_response.is_closed:
-            self._streaming_response.close()
-            self._streaming_response = None
-
-    def _process_incoming(self, resp: ChatResponse, assistant_index: int) -> None:
-        """Update local client-side state based on *resp*."""
-        if resp.type is ChatResponseType.STATE_UPDATE:
-            self.server_state = resp.content
-        elif resp.type is ChatResponseType.CONVERSATION_ID:
-            self.conversation_id = resp.content
-        elif resp.type is ChatResponseType.MESSAGE_ID:
-            pass
-        elif resp.type is ChatResponseType.TEXT:
-            assistant_msg = self.history[assistant_index]
-            if isinstance(resp.content, str):
-                assistant_msg.content += resp.content
-        elif resp.type is ChatResponseType.REFERENCE:
-            pass
+        """Stateless proxy to :pyclass:`Conversation.send_message`."""
+        conv = self.new_conversation()
+        yield from conv.send_message(message, context=context)
